requirements.txt
certifi==2017.4.17
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.1
django-braces==1.11.0
django-cors-middleware==1.3.1
django-oauth-toolkit==0.12.0
django-rest-framework-social-oauth2==1.0.8
djangorestframework==3.6.3
gunicorn==19.7.1
idna==2.5
oauthlib==2.0.1
olefile==0.44
Pillow==4.1.1
psycopg2==2.7.1
PyJWT==1.5.2
python3-openid==3.1.0
pytz==2017.2
requests==2.18.1
requests-oauthlib==0.8.0
six==1.10.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
urllib3==1.21.1
